Progressive rendering (WIP)
We are now packing all the data computed during a pass (in the accumulation buffer) into a texture that will be displayed onscreen using our shader.
The same sample accumulation method is also used when we export our result to a .PPM file.
While this currently work well when exporting, there still are problems when we display the texture, as the result does not seem to converge and shows vertical banding lines on the texture. Might not be using the right texture format or the data I am sending to the texture using glTexSubImage2D are not good...