LumenEngine rendering features, color space in shader
- Here it is ! Everything from LumenEngine has been (kinda) correctly ported to Tracer !
Just added everything rendering related in a nearly straight port fashion. Aside from this part, we have solid foundations for the future of the engine.
The reason the rendering part was strictly ported is because a lot of things are incorrect and/or lazy. The next proper step is now to... ditch everything about rendering/material that was just ported in this commit and implement a proper physically based BSDF (just wanted to have a complete port as a milestone, but in itself, it was kinda useless hehe).
- Moved the color space conversion process to the OpenGL shader stage. Was hesitant about porting the tonemapping operator from GLEngine, but since they are not exactly the kind of tonemapping used in a path tracer, I didn't do it.
- Added some more operator overloading in Vector.
- Added a beautiful(/stupid) trick in order to flip the OpenGL texture correctly in the shader.