Pixar USD support for scene description
- Quite a big addition this time, as I added a relatively simple support for the Pixar USD format. USD is a very interesting scene description format to work with, and its API surprisingly flexible regarding the way you can traverse the Scene Graph and retrieve informations.
The support is not, and cannot be perfect, as I am not 100% trying to map the "Pixar way" of doing things. I am for example not using the Hydra Engine for rendering, and what Tracer understand as a sphere primitive is not necessarily what Pixar and USD understand as such. But still!
- As a consequence, reworked the way I am loading a scene with a new SceneManager, that will use the same data structures to store scene informations, whether I am using a XML or USD scene file.
- Added a FindUSD CMake script that ease the building process. It might be solely Linux compatible however...
- Did a bit of cleaning in the includes in order to make interdependency harder to happen when coding.