Keep supporting USD
Contrary to what I stated in my last commit, I am still going to continue supporting USD in Tracer, making it in fact the underlying data format of the project, instead of going the custom JSON/SceneGraph way. I think it makes sense for me, given that I do work with USD at a professional level in the VFX industry, to support this new industry standard in the future.
The objective this time will be to have a proper support of USD, and not use it as kind of a data dumping format as I was doing up until now.
Embree/TBB supporting is still coming, though, with potentially triangle mesh rendering at the same time as it will take care of the BVH for me :)